Move AT inspection-related stuff into its own package
