+++ src/main/kotlin/platform/mcp/at/inspections/AtInspectionSuppressor.kt
@@ -18,8 +18,10 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.demonwav.mcdev.platform.mcp.at
+package com.demonwav.mcdev.platform.mcp.at.inspections
 
+import com.demonwav.mcdev.platform.mcp.at.AtElementFactory
+import com.demonwav.mcdev.platform.mcp.at.AtFile
 import com.demonwav.mcdev.platform.mcp.at.gen.psi.AtEntry
 import com.intellij.codeInspection.InspectionSuppressor
 import com.intellij.codeInspection.LocalQuickFixOnPsiElement
